--- a/frontend/packages/cli/package.json
+++ b/frontend/packages/cli/package.json
@@ -56,7 +56,7 @@
     "build": "pnpm run '/^build:.*/'",
     "build:cli": "rollup -c  && pnpm run cp:prism",
     "build:vite": "vite build",
-    "command:build": "node ./dist-cli/bin/cli.js erd build --input fixtures/input.schema.rb --format schemarb",
+    "command:build": "node ./dist-cli/bin/cli.js erd build --input https://github.com/mastodon/mastodon/blob/v4.3.2/db/schema.rb --format schemarb",
     "cp:prism": "cp ../db-structure/node_modules/@ruby/prism/src/prism.wasm ./dist-cli/bin/prism.wasm",
     "dev": "pnpm command:build && cp dist/schema.json public/ && pnpm run '/^dev:.*/'",
     "dev:app": "vite --port 5173",