--- a/frontend/packages/erd-core/src/components/Search/SearchOverlay.module.css
+++ b/frontend/packages/erd-core/src/components/Search/SearchOverlay.module.css
@@ -1,0 +1,112 @@
+.hiddenButton {
+  display: none;
+}
+
+.overlay {
+  position: fixed;
+  inset: 0;
+  background-color: rgba(0, 0, 0, 0.3);
+  backdrop-filter: blur(4px);
+  z-index: 50;
+}
+
+.content {
+  position: fixed;
+  top: 10vh;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 98vw;
+  max-width: 640px;
+  background-color: var(--fd-popover, white);
+  color: var(--fd-popover-foreground, black);
+  border-radius: 0.5rem;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+  z-index: 50;
+}
+
+.searchContainer {
+  display: flex;
+  flex-direction: column;
+  padding: 1rem;
+}
+
+.inputWrapper {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.input {
+  flex: 1;
+  padding: 0.75rem;
+  border: none;
+  background: transparent;
+  font-size: 1rem;
+  outline: none;
+  color: inherit;
+}
+
+.input::placeholder {
+  color: var(--fd-muted-foreground, #9ca3af);
+}
+
+.closeButton {
+  padding: 0.375rem 0.75rem;
+  font-size: 0.75rem;
+  border: 1px solid #d1d5db;
+  border-radius: 0.375rem;
+  background-color: transparent;
+  cursor: pointer;
+}
+
+.closeButton:hover {
+  background-color: #f3f4f6;
+}
+
+.resultsContainer {
+  display: flex;
+  flex-direction: column;
+  max-height: 460px;
+  overflow-y: auto;
+  border-top: 1px solid #e5e7eb;
+  padding: 0.5rem;
+}
+
+.resultItem {
+  display: flex;
+  align-items: center;
+  gap: 0.625rem;
+  min-height: 2.5rem;
+  padding: 0.5rem;
+  border-radius: 0.5rem;
+  text-align: left;
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+}
+
+.resultItem:hover {
+  background-color: var(--fd-accent, #e0f2fe);
+  color: var(--fd-accent-foreground, #1e3a8a);
+}
+
+.resultText {
+  flex: 1;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.icon {
+  width: 1rem;
+  height: 1rem;
+  color: var(--fd-muted-foreground, #9ca3af);
+}
+
+.indentLine {
+  margin-left: 0.5rem; /* ms-2 */
+  height: 100%;
+  min-height: 2.5rem; /* min-h-10 */
+  width: 1px;
+  background-color: var(--fd-border, #e5e7eb);
+}