--- a/frontend/packages/erd-core/src/components/Search/index.ts
+++ b/frontend/packages/erd-core/src/components/Search/index.ts
@@ -1,0 +1,1 @@
+export * from './SearchOverlay'