## Summary
<!-- Briefly describe the changes and the purpose of the PR. -->

PoC cmd-k search

- performance check
    - Orama
    - fuse

https://liam-erd-sample-cwsd0n63y-route-06-core.vercel.app/

## Related Issue
<!-- Mention the related issue number if applicable. -->

https://github.com/liam-hq/liam/issues/507

## Changes
<!-- List the main changes made in this PR. -->

## Testing
<!-- Briefly describe the testing steps or results. -->

## Other Information
<!-- Add any other relevant information for the reviewer. -->

### Comparison Summary

I compared fuse.js and orama for client-side search functionality. The key findings are as follows:

- **fuse.js** provides efficient fuzzy search capabilities out-of-the-box, without needing additional configuration for stemming or indexing.
- **Orama** excels in search speed due to pre-built inverted indexes and offers O(1) search performance, while fuse.js operates with O(n) complexity.
- Despite the theoretical performance differences, practical benchmarks on a project of Mastodon’s scale revealed that both libraries deliver comparable performance, with search times ranging from 0.1ms to 1.0ms.
- Orama’s pre-indexing time was approximately 30ms-100ms, initiated upon loading the application. fuse.js does not require pre-indexing, leading to simpler integration and faster initial load times.
- Both libraries successfully handled partial matches during live typing. However, Orama struggled with exact matches for incorrectly typed queries (e.g., "expire_at" instead of "expires_at").



### Comparison Detail

> Orama struggled with exact matches for incorrectly typed queries (e.g., "expire_at" instead of "expires_at").


https://github.com/user-attachments/assets/d34966db-96c9-4402-895a-b9965e1615af

------

> - search times ranging from 0.1ms to 1.0ms.

Both performances are excellent. 🚀 

<img width="986" alt="2025-02-05 19 33 19" src="https://github.com/user-attachments/assets/695890a8-c382-4772-b5d1-6e3baa1efa2d" />


------


> - Orama’s pre-indexing time was approximately 30-100ms, initiated upon loading the application. 




<img width="978" alt="2025-02-05 19 32 44" src="https://github.com/user-attachments/assets/077ca995-9758-4f46-8487-c4472554852c" />