@hoshinotsuyoshi 2025-02-05
ADR: https://github.com/liam-hq/liam/pull/675